Change the FactorInversion model, and also

+++ src/main/java/mx/com/innovating/cloud/data/models/FactorInversion.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class FactorInversion {
 
 	private Integer idhidrocarburo;
-	private Integer idoportunidadobjetivo;
+	private Long idoportunidadobjetivo;
 	private String hidrocarburo;
 	private Double pce;
 	private Double factorAceite;
